feat(frontend): add entity to collection using global state
