+++ apps/frontend/app/routes/_dashboard.media.groups.item.$id._index.tsx
@@ -10,7 +10,6 @@ import {
 	Text,
 	Title,
 } from "@mantine/core";
-import { useDisclosure } from "@mantine/hooks";
 import { unstable_defineLoader } from "@remix-run/node";
 import type { MetaArgs_SingleFetch } from "@remix-run/react";
 import { useLoaderData } from "@remix-run/react";
@@ -27,10 +26,7 @@ import {
 import invariant from "tiny-invariant";
 import { z } from "zod";
 import { zx } from "zodix";
-import {
-	AddEntityToCollectionModal,
-	MediaDetailsLayout,
-} from "~/components/common";
+import { MediaDetailsLayout } from "~/components/common";
 import {
 	DisplayCollection,
 	MediaIsPartial,
@@ -40,7 +36,7 @@ import {
 	ToggleMediaMonitorMenuItem,
 } from "~/components/media";
 import { useUserPreferences } from "~/lib/hooks";
-import { useReviewEntity } from "~/lib/state/media";
+import { useAddEntityToCollection, useReviewEntity } from "~/lib/state/media";
 import {
 	getAuthorizationHeader,
 	serverGqlService,
@@ -82,11 +78,8 @@ export const meta = ({ data }: MetaArgs_SingleFetch<typeof loader>) => {
 export default function Page() {
 	const loaderData = useLoaderData<typeof loader>();
 	const userPreferences = useUserPreferences();
-	const [
-		collectionModalOpened,
-		{ open: collectionModalOpen, close: collectionModalClose },
-	] = useDisclosure(false);
 	const [_r, setEntityToReview] = useReviewEntity();
+	const [_a, setAddEntityToCollectionData] = useAddEntityToCollection();
 
 	return (
 		<Container>
@@ -168,18 +161,21 @@ export default function Page() {
 								>
 									Post a review
 								</Button>
-								<Button variant="outline" onClick={collectionModalOpen}>
+								<Button
+									variant="outline"
+									onClick={() => {
+										setAddEntityToCollectionData({
+											entityId: loaderData.metadataGroupId,
+											entityLot: EntityLot.MetadataGroup,
+											alreadyInCollections:
+												loaderData.userMetadataGroupDetails.collections.map(
+													(c) => c.id,
+												),
+										});
+									}}
+								>
 									Add to collection
 								</Button>
-								<AddEntityToCollectionModal
-									onClose={collectionModalClose}
-									opened={collectionModalOpened}
-									entityId={loaderData.metadataGroupId.toString()}
-									entityLot={EntityLot.MetadataGroup}
-									alreadyInCollections={loaderData.userMetadataGroupDetails.collections.map(
-										(c) => c.id,
-									)}
-								/>
 								<Menu shadow="md">
 									<Menu.Target>
 										<Button variant="outline">More actions</Button>
